- all the view almost finish, there is only the pop ups functions to add - only the upload ImageLoader works, so still a lot left - change the singleton to make a consistent profile image - also logout only on SettingsActivity.kt
