+++ app/src/main/AndroidManifest.xml
@@ -4,10 +4,10 @@
 
     <uses-permission android:name="android.permission.INTERNET" />
     <uses-permission android:name="android.permission.ACCESS_FINE_LOCATION" />
-    <uses-permission android:name="android.permission.ACCESS_FINE_LOCATION" />
     <uses-permission android:name="android.permission.ACCESS_COARSE_LOCATION" />
 
     <application
+        android:name=".CustomApplication"
         android:allowBackup="true"
         android:dataExtractionRules="@xml/data_extraction_rules"
         android:fullBackupContent="@xml/backup_rules"
@@ -17,18 +17,16 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.InteractiveMapUniandes"
         android:usesCleartextTraffic="true"
-        tools:targetApi="31"
-        android:name=".CustomApplication">
+        tools:targetApi="31">
         <activity
-            android:name="interactivemapuniandes.ui.SettingsActivity"
+            android:name=".ui.SettingsActivity"
             android:exported="false" />
         <activity
             android:name=".ui.SignupActivity"
             android:exported="false" />
         <activity
             android:name=".ui.RouteActivity"
-            android:exported="false"
-            tools:ignore="MissingClass" />
+            android:exported="false" />
 
         <meta-data
             android:name="com.google.android.geo.API_KEY"
