+++ app/src/main/res/layout/activity_settings.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/background_color"
-    tools:context="interactivemapuniandes.ui.SettingsActivity">
+    tools:context="com.uniandes.interactivemapuniandes.ui.SettingsActivity">
 
     <!-- Settings Header -->
     <androidx.constraintlayout.widget.ConstraintLayout
@@ -235,14 +235,22 @@
                     app:layout_constraintStart_toEndOf="@id/dark_mode_image_settings"
                     app:layout_constraintTop_toTopOf="@id/dark_mode_image_settings" />
 
-                <TextView
+                <androidx.constraintlayout.widget.ConstraintLayout
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
-                    android:text="Off"
-                    android:textColor="@color/settings_secondary_text"
                     app:layout_constraintBottom_toBottomOf="parent"
                     app:layout_constraintEnd_toEndOf="parent"
-                    app:layout_constraintTop_toTopOf="parent" />
+                    app:layout_constraintTop_toTopOf="parent">
+
+                    <com.google.android.material.materialswitch.MaterialSwitch
+                        android:layout_width="wrap_content"
+                        android:layout_height="wrap_content"
+                        android:checked="true"
+                        app:layout_constraintBottom_toBottomOf="parent"
+                        app:layout_constraintEnd_toEndOf="parent"
+                        app:layout_constraintTop_toTopOf="parent"/>
+
+                </androidx.constraintlayout.widget.ConstraintLayout>
 
             </androidx.constraintlayout.widget.ConstraintLayout>
