+++ app/src/main/res/layout/activity_signup.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     android:background="@color/background_color"
     android:fillViewport="true"
-    tools:context=".ui.SignupActivity">
+    tools:context="com.uniandes.interactivemapuniandes.ui.SignupActivity">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
